python/requirements.txt
pandas==1.3.4
matplotlib
Jinja2
scikit-learn
python-dotenv